restructuring tests; removing helpers
