+++ ownables/potion/src/contract.rs
@@ -224,7 +224,7 @@ pub fn try_set_bridge(info: MessageInfo, deps: DepsMut, addr: Option<Addr>) -> R
     Ok(resp)
 }
 
-pub fn query(deps: Deps, msg: QueryMsg) -> StdResult<Binary> {
+pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
     match msg {
         QueryMsg::GetOwnableConfig {} => query_ownable_config(deps),
         QueryMsg::GetOwnableMetadata {} => query_ownable_metadata(deps),
@@ -266,186 +266,4 @@ fn query_ownable_metadata(deps: Deps) -> StdResult<Binary> {
         animation_url: config.animation_url,
         youtube_url: config.youtube_url,
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use std::marker::PhantomData;
-    use super::*;
-    use cosmwasm_std::testing::{mock_env, mock_info};
-    use cosmwasm_std::{Addr, OwnedDeps, MemoryStorage, Response, MessageInfo};
-    use crate::utils::{EmptyApi, EmptyQuerier};
-
-    struct CommonTest {
-        deps: OwnedDeps<MemoryStorage, EmptyApi, EmptyQuerier>,
-        info: MessageInfo,
-        res: Response,
-    }
-    fn setup_test() -> CommonTest {
-        let mut deps = OwnedDeps {
-            storage: MemoryStorage::default(),
-            api: EmptyApi::default(),
-            querier: EmptyQuerier::default(),
-            custom_query_type: PhantomData,
-        };
-        let info = mock_info("sender-1", &[]);
-
-        let msg = InstantiateMsg {
-            ownable_id: "2bJ69cFXzS8AJTcCmzjc9oeHZmBrmMVUr8svJ1mTGpho9izYrbZjrMr9q1YwvY".to_string(),
-            image: None,
-            image_data: None,
-            external_url: None,
-            description: Some("ownable to consume".to_string()),
-            name: Some("Potion".to_string()),
-            background_color: None,
-            animation_url: None,
-            youtube_url: None
-        };
-
-        let res: Response = instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        CommonTest {
-            deps,
-            info,
-            res,
-        }
-    }
-
-    #[test]
-    fn test_initialize() {
-
-        let CommonTest {
-            deps: _,
-            info: _,
-            res,
-        } = setup_test();
-
-        assert_eq!(0, res.messages.len());
-        assert_eq!(res.attributes.get(0).unwrap().value, "instantiate".to_string());
-        assert_eq!(res.attributes.get(1).unwrap().value, "sender-1".to_string());
-        assert_eq!(res.attributes.get(2).unwrap().value, "sender-1".to_string());
-        assert!(res.attributes.get(3).unwrap().value.starts_with("#"));
-        assert_eq!(res.attributes.get(3).unwrap().value.len(), 7);
-        assert_eq!(res.attributes.get(4).unwrap().value, "100".to_string());
-    }
-
-    #[test]
-    fn test_consume() {
-        let CommonTest {
-            mut deps,
-            info,
-            res: _,
-        } = setup_test();
-        let deps = deps.as_mut();
-
-        let msg = ExecuteMsg::Consume { amount: 50 };
-        let res: Response = execute(deps, mock_env(), info, msg).unwrap();
-
-        assert_eq!(0, res.messages.len());
-        assert_eq!(res.attributes.get(0).unwrap().value, "try_consume".to_string());
-        assert_eq!(res.attributes.get(1).unwrap().value, "50".to_string());
-    }
-
-    #[test]
-    fn test_consume_unauthorized() {
-        let CommonTest {
-            mut deps,
-            mut info,
-            res: _,
-        } = setup_test();
-        let deps = deps.as_mut();
-        info.sender = Addr::unchecked("not-the-owner".to_string());
-        let msg = ExecuteMsg::Consume { amount: 50 };
-
-        let err: ContractError = execute(deps, mock_env(), info, msg)
-            .unwrap_err();
-
-        let _expected_err_val = "Unauthorized consumption attempt".to_string();
-        assert!(matches!(err, ContractError::Unauthorized { val: _expected_err_val, }));
-    }
-
-    #[test]
-    fn test_overconsume() {
-        let CommonTest {
-            mut deps,
-            info,
-            res: _,
-        } = setup_test();
-        let deps = deps.as_mut();
-        let msg = ExecuteMsg::Consume { amount: 150 };
-
-        let err: ContractError = execute(deps, mock_env(), info, msg)
-            .unwrap_err();
-
-        let _expected_err_val = "attempt to consume more than possible".to_string();
-        assert!(matches!(err, ContractError::CustomError { val: _expected_err_val }));
-    }
-
-    #[test]
-    fn test_transfer() {
-        let CommonTest {
-            mut deps,
-            info,
-            res: _,
-        } = setup_test();
-        let deps = deps.as_mut();
-
-        let msg = ExecuteMsg::Transfer { to: Addr::unchecked("other-owner-1") };
-
-        let res: Response = execute(deps, mock_env(), info, msg).unwrap();
-
-        assert_eq!(res.attributes.get(0).unwrap().value, "try_transfer".to_string());
-        assert_eq!(res.attributes.get(1).unwrap().value, "other-owner-1".to_string());
-    }
-
-    #[test]
-    fn test_transfer_unauthorized() {
-        let CommonTest {
-            mut deps,
-            mut info,
-            res: _,
-        } = setup_test();
-        let deps = deps.as_mut();
-        info.sender = Addr::unchecked("not-the-owner".to_string());
-        let msg = ExecuteMsg::Transfer { to: Addr::unchecked("not-the-owner") };
-
-        let err: ContractError = execute(deps, mock_env(), info, msg)
-            .unwrap_err();
-
-        let _expected_err_val = "Unauthorized transfer attempt".to_string();
-        assert!(matches!(err, ContractError::Unauthorized { val: _expected_err_val }));
-    }
-
-    #[test]
-    fn test_query_config() {
-        let CommonTest {
-            deps,
-            info: _,
-            res: _,
-        } = setup_test();
-
-        let msg = QueryMsg::GetOwnableConfig {};
-        let resp: Binary = query(deps.as_ref(), msg).unwrap();
-        let json: String = "{\"owner\":\"sender-1\",\"issuer\":\"sender-1\",\"current_amount\":100,\"max_capacity\":100,\"color\":\"#11D539\"}".to_string();
-        let expected_binary = Binary::from(json.as_bytes());
-
-        assert_eq!(resp, expected_binary);
-    }
-
-    #[test]
-    fn test_query_metadata() {
-        let CommonTest {
-            deps,
-            info: _,
-            res: _,
-        } = setup_test();
-
-        let msg = QueryMsg::GetOwnableMetadata {};
-        let resp: Binary = query(deps.as_ref(), msg).unwrap();
-        let json: String = "{\"image\":null,\"image_data\":null,\"external_url\":null,\"description\":\"Ownable potion that can be consumed\",\"name\":\"Potion\",\"background_color\":null,\"animation_url\":null,\"youtube_url\":null}".to_string();
-        let expected_binary = Binary::from(json.as_bytes());
-
-        assert_eq!(resp, expected_binary);
-    }
-
 }
